fix: Trailing Stop adjusted_ts 타입 충돌 수정 (int → float)
문제:
- round() * 10 연산이 int 반환
- number_input 파라미터들은 float 타입
- StreamlitMixedNumericTypesError 발생

해결:
- float() 명시적 타입 변환 추가
- line 585: adjusted_ts = float(max(...))

+++ pages/set_buy_sell_conditions.py
@@ -582,7 +582,7 @@ with st.expander("⭐ 핵심 전략 조건", expanded=True):
 
                 # ✅ 기존 값을 10의 배수로 자동 반올림 (범위: 10~90)
                 current_ts = st.session_state.get("trailing_stop_threshold_pct", 10.0)
-                adjusted_ts = max(10.0, min(90.0, round(current_ts / 10) * 10))
+                adjusted_ts = float(max(10.0, min(90.0, round(current_ts / 10) * 10)))
 
                 ts_threshold = st.number_input(
                     "수익 하락 허용 (%)",
